+++ config/environments/development.rb
@@ -6,6 +6,10 @@ Rails.application.configure do
   # In the development environment your application's code is reloaded any time
   # it changes. This slows down response time but is perfect for development
   # since you don't have to restart the web server when you make code changes.
+
+  #add a line 
+  config.autoload_paths << Rails.root.join('app', 'controllers', 'api', 'v1')
+ # 
   config.cache_classes = false
 
   # Do not eager load code on boot.
